requirements.txt

streamlit==1.28.0
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.20
yolov5==7.0.13
Pillow==10.0.0
opencv-python==4.8.0.76
matplotlib==3.7.2
numpy==1.24.3
pandas==1.5.3